Test.QuickCheck.Poly

Description

Types to help with testing polymorphic properties.

Types A, B and C are newtype wrappers around Integer that implement Eq, Show, Arbitrary and CoArbitrary. Types OrdA, OrdB and OrdC also implement Ord and Num.

See also Test.QuickCheck.All for an automatic way of testing polymorphic properties.

Documentation

newtype A

Constructors

A

Fields

unA :: Integer

newtype B

Constructors

B

Fields

unB :: Integer

newtype C

Constructors

C

Fields

unC :: Integer

newtype OrdA

Constructors

OrdA

Fields

unOrdA :: Integer

newtype OrdB

Constructors

OrdB

Fields

unOrdB :: Integer

newtype OrdC

Constructors

OrdC

Fields

unOrdC :: Integer