Test.QuickCheck

Running tests

quickCheck :: Testable prop => prop -> IO ()

Tests a property and prints the results to stdout.

data Args

Args specifies arguments to the QuickCheck driver

Constructors

Args

Fields

replay :: Maybe (StdGen, Int)

should we replay a previous test?

maxSuccess :: Int

maximum number of successful tests before succeeding

maxDiscardRatio :: Int

maximum number of discarded tests per successful test before giving up

maxSize :: Int

size to use for the biggest test cases

chatty :: Bool

whether to print anything

data Result

Result represents the test result

Constructors

Success

Fields

numTests :: Int

number of successful tests performed

labels :: [(String, Int)]

labels and frequencies found during all tests

output :: String

printed output

GaveUp

Fields

numTests :: Int

number of successful tests performed

labels :: [(String, Int)]

labels and frequencies found during all tests

output :: String

printed output

Failure

Fields

numTests :: Int

number of successful tests performed

numShrinks :: Int

number of successful shrinking steps performed

usedSeed :: StdGen

what seed was used

usedSize :: Int

what was the test size

reason :: String

what was the reason

interrupted :: Bool

did the user press ctrl-C?

labels :: [(String, Int)]

labels and frequencies found during all tests

output :: String

printed output

NoExpectedFailure

Fields

numTests :: Int

number of successful tests performed

labels :: [(String, Int)]

labels and frequencies found during all tests

output :: String

printed output

stdArgs :: Args

stdArgs are the default test arguments used

quickCheckWith :: Testable prop => Args -> prop -> IO ()

Tests a property, using test arguments, and prints the results to stdout.

quickCheckWithResult :: Testable prop => Args -> prop -> IO Result

Tests a property, using test arguments, produces a test result, and prints the results to stdout.

quickCheckResult :: Testable prop => prop -> IO Result

Tests a property, produces a test result, and prints the results to stdout.

Running tests verbosely

verboseCheck :: Testable prop => prop -> IO ()

Tests a property and prints the results and all test cases generated to stdout. This is just a convenience function that means the same as quickCheck . verbose.

verboseCheckWith :: Testable prop => Args -> prop -> IO ()

Tests a property, using test arguments, and prints the results and all test cases generated to stdout. This is just a convenience function that combines quickCheckWith and verbose.

verboseCheckWithResult :: Testable prop => Args -> prop -> IO Result

Tests a property, using test arguments, produces a test result, and prints the results and all test cases generated to stdout. This is just a convenience function that combines quickCheckWithResult and verbose.

verboseCheckResult :: Testable prop => prop -> IO Result

Tests a property, produces a test result, and prints the results and all test cases generated to stdout. This is just a convenience function that combines quickCheckResult and verbose.

verbose :: Testable prop => prop -> Property

Prints out the generated testcase every time the property is tested, like verboseCheck from QuickCheck 1. Only variables quantified over inside the verbose are printed.

Random generation

data Gen a

Generator combinators

sized :: (Int -> Gen a) -> Gen a

Used to construct generators that depend on the size parameter.

resize :: Int -> Gen a -> Gen a

Overrides the size parameter. Returns a generator which uses the given size instead of the runtime-size parameter.

choose :: Random a => (a, a) -> Gen a

Generates a random element in the given inclusive range.

promote :: Monad m => m (Gen a) -> Gen (m a)

Promotes a monadic generator to a generator of monadic values.

suchThat :: Gen a -> (a -> Bool) -> Gen a

Generates a value that satisfies a predicate.

suchThatMaybe :: Gen a -> (a -> Bool) -> Gen (Maybe a)

Tries to generate a value that satisfies a predicate.

oneof :: [Gen a] -> Gen a

Randomly uses one of the given generators. The input list must be non-empty.

frequency :: [(Int, Gen a)] -> Gen a

Chooses one of the given generators, with a weighted random distribution. The input list must be non-empty.

elements :: [a] -> Gen a

Generates one of the given values. The input list must be non-empty.

growingElements :: [a] -> Gen a

Takes a list of elements of increasing size, and chooses among an initial segment of the list. The size of this initial segment increases with the size parameter. The input list must be non-empty.

listOf :: Gen a -> Gen [a]

Generates a list of random length. The maximum length depends on the size parameter.

listOf1 :: Gen a -> Gen [a]

Generates a non-empty list of random length. The maximum length depends on the size parameter.

vectorOf :: Int -> Gen a -> Gen [a]

Generates a list of the given length.

Generators which use Arbitrary

vector :: Arbitrary a => Int -> Gen [a]

Generates a list of a given length.

orderedList :: (Ord a, Arbitrary a) => Gen [a]

Generates an ordered list of a given length.

Generator debugging

sample :: Show a => Gen a -> IO ()

Generates some example values and prints them to stdout.

sample' :: Gen a -> IO [a]

Generates some example values.

Arbitrary and CoArbitrary classes

class Arbitrary a where

Random generation and shrinking of values.

Minimal complete definition

Nothing

Methods

arbitrary :: Gen a

A generator for values of the given type.

shrink :: a -> [a]

Produces a (possibly) empty list of all the possible immediate shrinks of the given value.

class CoArbitrary a where

Used for random generation of functions.

Methods

coarbitrary :: a -> Gen c -> Gen c

Used to generate a function of type a -> c. The implementation should use the first argument to perturb the random generator given as the second argument. the returned generator is then used to generate the function result. You can often use variant and >< to implement coarbitrary.

Helper functions for implementing arbitrary

arbitrarySizedIntegral :: Num a => Gen a

Generates an integral number. The number can be positive or negative and its maximum absolute value depends on the size parameter.

arbitrarySizedFractional :: Fractional a => Gen a

Generates a fractional number. The number can be positive or negative and its maximum absolute value depends on the size parameter.

arbitrarySizedBoundedIntegral :: (Bounded a, Integral a) => Gen a

Generates an integral number from a bounded domain. The number is chosen from the entire range of the type, but small numbers are generated more often than big numbers. Inspired by demands from Phil Wadler.

arbitraryBoundedIntegral :: (Bounded a, Integral a) => Gen a

Generates an integral number. The number is chosen uniformly from the entire range of the type. You may want to use arbitrarySizedBoundedIntegral instead.

arbitraryBoundedRandom :: (Bounded a, Random a) => Gen a

Generates an element of a bounded type. The element is chosen from the entire range of the type.

arbitraryBoundedEnum :: (Bounded a, Enum a) => Gen a

Generates an element of a bounded enumeration.

coarbitraryEnum :: Enum a => a -> Gen b -> Gen b

A coarbitrary implementation for enums.

Helper functions for implementing shrink

shrinkNothing :: a -> [a]

Returns no shrinking alternatives.

shrinkIntegral :: Integral a => a -> [a]

Shrink an integral number.

shrinkRealFrac :: RealFrac a => a -> [a]

Shrink a fraction.

Helper functions for implementing coarbitrary

variant :: Integral n => n -> Gen a -> Gen a

Modifies a generator using an integer seed.

(><) :: (Gen a -> Gen a) -> (Gen a -> Gen a) -> Gen a -> Gen a

Combine two generator perturbing functions, for example the results of calls to variant or coarbitrary.

coarbitraryIntegral :: Integral a => a -> Gen b -> Gen b

A coarbitrary implementation for integral numbers.

coarbitraryReal :: Real a => a -> Gen b -> Gen b

A coarbitrary implementation for real numbers.

coarbitraryShow :: Show a => a -> Gen b -> Gen b

coarbitrary helper for lazy people :-).

Type-level modifiers for changing generator behavior

newtype Blind a

Blind x: as x, but x does not have to be in the Show class.

Constructors

Blind a

newtype Fixed a

Fixed x: as x, but will not be shrunk.

Constructors

Fixed a

newtype OrderedList a

Ordered xs: guarantees that xs is ordered.

Constructors

Ordered

Fields

getOrdered :: [a]

newtype NonEmptyList a

NonEmpty xs: guarantees that xs is non-empty.

Constructors

NonEmpty

Fields

getNonEmpty :: [a]

newtype Positive a

Positive x: guarantees that x > 0.

Constructors

Positive

Fields

getPositive :: a

newtype NonZero a

NonZero x: guarantees that x /= 0.

Constructors

NonZero

Fields

getNonZero :: a

newtype NonNegative a

NonNegative x: guarantees that x >= 0.

Constructors

NonNegative

Fields

getNonNegative :: a

data Smart a

Smart _ x: tries a different order when shrinking.

Constructors

Smart Int a

newtype Shrink2 a

Shrink2 x: allows 2 shrinking steps at the same time when shrinking x

Constructors

Shrink2 a

data Shrinking s a

Shrinking _ x: allows for maintaining a state during shrinking.

Constructors

Shrinking s a

class ShrinkState s a where

Methods

shrinkInit :: a -> s

shrinkState :: a -> s -> [(a, s)]

Properties

type Property = Gen Prop

data Prop

class Testable prop where

The class of things which can be tested, i.e. turned into a property.

Minimal complete definition

property

Methods

property :: prop -> Property

exhaustive :: prop -> Bool

Property combinators

mapSize :: Testable prop => (Int -> Int) -> prop -> Property

Changes the maximum test case size for a property.

shrinking

Arguments

:: Testable prop

=> (a -> [a])
shrink-like function.

-> a
The original argument

-> (a -> prop)

-> Property

Shrinks the argument to property if it fails. Shrinking is done automatically for most types. This is only needed when you want to override the default behavior.

(==>) :: Testable prop => Bool -> prop -> Property infixr 0

Implication for properties: The resulting property holds if the first argument is False (in which case the test case is discarded), or if the given property holds.

discard :: a

A special exception that makes QuickCheck discard the test case. Normally you should use ==>, but if for some reason this isn't possible (e.g. you are deep inside a generator), use discard instead.

forAll :: (Show a, Testable prop) => Gen a -> (a -> prop) -> Property

Explicit universal quantification: uses an explicitly given test case generator.

forAllShrink :: (Show a, Testable prop) => Gen a -> (a -> [a]) -> (a -> prop) -> Property

Like forAll, but tries to shrink the argument for failing test cases.

Experimental combinators for conjunction and disjunction

(.&.) :: (Testable prop1, Testable prop2) => prop1 -> prop2 -> Property infixr 1

Nondeterministic choice: p1 .&. p2 picks randomly one of p1 and p2 to test. If you test the property 100 times it makes 100 random choices.

(.&&.) :: (Testable prop1, Testable prop2) => prop1 -> prop2 -> Property infixr 1

Conjunction: p1 .&&. p2 passes if both p1 and p2 pass.

conjoin :: Testable prop => [prop] -> Property

Take the conjunction of several properties.

(.||.) :: (Testable prop1, Testable prop2) => prop1 -> prop2 -> Property infixr 1

Disjunction: p1 .||. p2 passes unless p1 and p2 simultaneously fail.

disjoin :: Testable prop => [prop] -> Property

Take the disjunction of several properties.

Handling failure

whenFail :: Testable prop => IO () -> prop -> Property

Performs an IO action after the last failure of a property.

printTestCase :: Testable prop => String -> prop -> Property

Prints a message to the terminal as part of the counterexample.

whenFail' :: Testable prop => IO () -> prop -> Property

Performs an IO action every time a property fails. Thus, if shrinking is done, this can be used to keep track of the failures along the way.

expectFailure :: Testable prop => prop -> Property

Modifies a property so that it is expected to fail for some test cases.

within :: Testable prop => Int -> prop -> Property

Considers a property failed if it does not complete within the given number of microseconds.

Test distribution

label :: Testable prop => String -> prop -> Property

Attaches a label to a property. This is used for reporting test case distribution.

collect :: (Show a, Testable prop) => a -> prop -> Property

Labels a property with a value:

collect x = label (show x)

classify

Arguments

:: Testable prop

=> Bool
True if the test case should be labelled.

-> String
Label.

-> prop

-> Property

Conditionally labels test case.

cover

Arguments

:: Testable prop

=> Bool
True if the test case belongs to the class.

-> Int
The required percentage (0-100) of test cases.

-> String
Label for the test case class.

-> prop

-> Property

Checks that at least the given proportion of the test cases belong to the given class.

once :: Testable prop => prop -> Property

Modifies a property so that it only will be tested once.

Text formatting

newtype Str

Constructors

MkStr String

ranges :: (Show a, Integral a) => a -> a -> Str