Test.QuickCheck.Gen

Description

Test case generation.

Generator type

newtype Gen a

A generator for values of type a.

Constructors

MkGen

Fields

unGen :: QCGen -> Int -> a

Run the generator on a particular seed. If you just want to get a random value out, consider using generate.

Primitive generator combinators

variant :: Integral n => n -> Gen a -> Gen a

Modifies a generator using an integer seed.

sized :: (Int -> Gen a) -> Gen a

Used to construct generators that depend on the size parameter.

resize :: Int -> Gen a -> Gen a

Overrides the size parameter. Returns a generator which uses the given size instead of the runtime-size parameter.

scale :: (Int -> Int) -> Gen a -> Gen a

Adjust the size parameter, by transforming it with the given function.

choose :: Random a => (a, a) -> Gen a

Generates a random element in the given inclusive range.

generate :: Gen a -> IO a

Run a generator. The size passed to the generator is always 30; if you want another size then you should explicitly use resize.

sample' :: Gen a -> IO [a]

Generates some example values.

sample :: Show a => Gen a -> IO ()

Generates some example values and prints them to stdout.

Common generator combinators

suchThat :: Gen a -> (a -> Bool) -> Gen a

Generates a value that satisfies a predicate.

suchThatMaybe :: Gen a -> (a -> Bool) -> Gen (Maybe a)

Tries to generate a value that satisfies a predicate.

oneof :: [Gen a] -> Gen a

Randomly uses one of the given generators. The input list must be non-empty.

frequency :: [(Int, Gen a)] -> Gen a

Chooses one of the given generators, with a weighted random distribution. The input list must be non-empty.

elements :: [a] -> Gen a

Generates one of the given values. The input list must be non-empty.

sublistOf :: [a] -> Gen [a]

Generates a random subsequence of the given list.

shuffle :: [a] -> Gen [a]

Generates a random permutation of the given list.

growingElements :: [a] -> Gen a

Takes a list of elements of increasing size, and chooses among an initial segment of the list. The size of this initial segment increases with the size parameter. The input list must be non-empty.

listOf :: Gen a -> Gen [a]

Generates a list of random length. The maximum length depends on the size parameter.

listOf1 :: Gen a -> Gen [a]

Generates a non-empty list of random length. The maximum length depends on the size parameter.

vectorOf :: Int -> Gen a -> Gen [a]

Generates a list of the given length.

infiniteListOf :: Gen a -> Gen [a]

Generates an infinite list.